+++ src/main/java/com/hp/hpl/jena/tdb/base/file/BlockAccessByteArray.java
@@ -86,7 +86,7 @@ public class BlockAccessByteArray implements BlockAccess
     }
 
     @Override
-    public boolean valid(int id)
+    public boolean valid(long id)
     {
         return ( id >= 0 && id < length ) ;
     }
